Navradio: disable low-pass filter when changing selected freq or leaving GPS slave mode
When the selected frequency is changed or when leaving GPS slave mode,
disable the low-pass filter applied to signal quality. Otherwise, the
following may happen for instance:
  - the active frequency corresponds to a navaid whose signal is well
    received;
  - user makes it so that the Morse code for the navaid ID can be heard;
  - user then changes the selected frequency to one that doesn't belong
    to a navaid that is close enough to be usable;
  - yet, as soon as the frequency is changed, the Morse code for the ID
    of the newly selected navaid (if any), even if it is way too far for
    its signal to be received, will be very clearly heard for about one
    second---and likely truncated.

This is because before this commit, after the frequency change, the
low-pass filter applied to signal quality made the code behave as if the
signal, supposedly coming from the new navaid, were still strong---which
of course doesn't correspond to physical reality. This fixes the bug
reported at [1].

[1] https://forum.flightgear.org/viewtopic.php?f=25&t=40890#p405708