Ticket #2770 : ATC/TowerController.cxx:203 Floods .fgfs.log
Disabled code that is meant for troubleshooting (DEBUG), but is categorized as BULK.

I did not change the category of this message log, because there seems to be another problem where the user-specified log level is being altered.  Disabled this message until a more thorough review of the logging code is made.